healthcare_backend/requirements.txt
python-multipart
fastapi[standard]
uvicorn
pydantic
python-dotenv
requests
pandas
openpyxl
mysql-connector-python